fix: PG identity sequence desync — silent message loss for all new sessions
Root cause: migrate_from_sqlite inserts rows with explicit IDs but
After migration, sequence value << max(id), causing every new insert
to collide on PK. ON CONFLICT DO NOTHING (untargeted) silently
swallowed PK collisions — zero error logs, zero messages stored.

Fixes:
- Target ON CONFLICT to specific constraint columns:
  conversation_messages → ON CONFLICT (message_uuid)
  conversation_events → ON CONFLICT (session_id, event_uuid)
  PK collisions now properly raise errors instead of silent drops
- Add PgMissionStore::fix_identity_sequences() startup health check
- Add post-migration sequence fix in migrate_from_sqlite
- Fix INT4/INT8 type mismatch in narration_cursors (retrospective.rs)

+++ crates/missiond-core/src/db/pg/retrospective.rs
@@ -278,7 +278,7 @@ impl RetrospectiveStore for PgMissionStore {
             row.get::<i64, _>("batch_index"),
             row.get::<String, _>("status"),
             row.get::<i64, _>("retry_count"),
-            row.get::<i64, _>("total_messages"),
+            row.get::<i32, _>("total_messages") as i64,
         ))
     }
